JAMES-2156 moved JPAConfiguration to backends/jpa
