+++ backends-common/jpa/src/main/java/org/apache/james/backends/jpa/JPAConfiguration.java
@@ -16,17 +16,17 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.modules.data;
+package org.apache.james.backends.jpa;
 
-import static org.apache.james.modules.data.JPAConfiguration.Credential.NO_CREDENTIAL;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.CUSTOM_DATASOURCE_PROPERTIES;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.CUSTOM_OPENJPA_PROPERTIES;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.NO_ATTACHMENT_STORAGE;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.NO_MAX_CONNECTIONS;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.NO_MULTITHREADED;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.NO_TEST_ON_BORROW;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.NO_VALIDATION_QUERY;
-import static org.apache.james.modules.data.JPAConfiguration.ReadyToBuild.NO_VALIDATION_QUERY_TIMEOUT_SEC;
+import static org.apache.james.backends.jpa.JPAConfiguration.Credential.NO_CREDENTIAL;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.CUSTOM_DATASOURCE_PROPERTIES;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.CUSTOM_OPENJPA_PROPERTIES;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.NO_ATTACHMENT_STORAGE;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.NO_MAX_CONNECTIONS;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.NO_MULTITHREADED;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.NO_TEST_ON_BORROW;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.NO_VALIDATION_QUERY;
+import static org.apache.james.backends.jpa.JPAConfiguration.ReadyToBuild.NO_VALIDATION_QUERY_TIMEOUT_SEC;
 
 import java.util.HashMap;
 import java.util.List;
@@ -41,21 +41,21 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 
 public class JPAConfiguration {
-    static final String JPA_CONNECTION_DRIVER_NAME = "openjpa.ConnectionDriverName";
-    static final String JPA_CONNECTION_USERNAME = "openjpa.ConnectionUserName";
-    static final String JPA_CONNECTION_PASSWORD = "openjpa.ConnectionPassword";
-    static final String JPA_CONNECTION_PROPERTIES = "openjpa.ConnectionProperties";
-    static final String JPA_CONNECTION_URL = "openjpa.ConnectionURL";
-    static final String JPA_MULTITHREADED = "openjpa.Multithreaded";
-    static List<String> DEFAULT_JPA_PROPERTIES = List.of(JPA_CONNECTION_DRIVER_NAME, JPA_CONNECTION_URL, JPA_MULTITHREADED, JPA_CONNECTION_USERNAME, JPA_CONNECTION_PASSWORD);
+    public static final String JPA_CONNECTION_DRIVER_NAME = "openjpa.ConnectionDriverName";
+    public static final String JPA_CONNECTION_USERNAME = "openjpa.ConnectionUserName";
+    public static final String JPA_CONNECTION_PASSWORD = "openjpa.ConnectionPassword";
+    public static final String JPA_CONNECTION_PROPERTIES = "openjpa.ConnectionProperties";
+    public static final String JPA_CONNECTION_URL = "openjpa.ConnectionURL";
+    public static final String JPA_MULTITHREADED = "openjpa.Multithreaded";
+    public static final List<String> DEFAULT_JPA_PROPERTIES = List.of(JPA_CONNECTION_DRIVER_NAME, JPA_CONNECTION_URL, JPA_MULTITHREADED, JPA_CONNECTION_USERNAME, JPA_CONNECTION_PASSWORD);
 
-    static final String DATASOURCE_TEST_ON_BORROW = "datasource.testOnBorrow";
-    static final String DATASOURCE_VALIDATION_QUERY_TIMEOUT_SEC = "datasource.validationQueryTimeoutSec";
-    static final String DATASOURCE_VALIDATION_QUERY = "datasource.validationQuery";
-    static final String DATASOURCE_MAX_TOTAL = "datasource.maxTotal";
-    static List<String> DEFAULT_DATASOURCE_PROPERTIES = List.of(DATASOURCE_TEST_ON_BORROW, DATASOURCE_VALIDATION_QUERY_TIMEOUT_SEC, DATASOURCE_VALIDATION_QUERY, DATASOURCE_MAX_TOTAL);
+    public static final String DATASOURCE_TEST_ON_BORROW = "datasource.testOnBorrow";
+    public static final String DATASOURCE_VALIDATION_QUERY_TIMEOUT_SEC = "datasource.validationQueryTimeoutSec";
+    public static final String DATASOURCE_VALIDATION_QUERY = "datasource.validationQuery";
+    public static final String DATASOURCE_MAX_TOTAL = "datasource.maxTotal";
+    public static final List<String> DEFAULT_DATASOURCE_PROPERTIES = List.of(DATASOURCE_TEST_ON_BORROW, DATASOURCE_VALIDATION_QUERY_TIMEOUT_SEC, DATASOURCE_VALIDATION_QUERY, DATASOURCE_MAX_TOTAL);
 
-    static final String ATTACHMENT_STORAGE = "attachmentStorage.enabled";
+    public static final String ATTACHMENT_STORAGE = "attachmentStorage.enabled";
 
     static {
     }
