+++ mailbox/jpa/src/main/resources/james-database.properties
@@ -44,3 +44,8 @@ openjpa.streaming=false
 # datasource.validationQueryTimeoutSec=2
 # This is different per database. See https://stackoverflow.com/questions/10684244/dbcp-validationquery-for-different-databases#10684260
 # datasource.validationQuery=select 1
+
+# Attachment storage
+# *WARNING*: Is not made to store large binary content (no more than 1 GB of data)
+# Optional, Allowed values are: true, false, defaults to false
+# attachmentStorage.enabled=false
