+++ server/apps/jpa-smtp-app/sample-configuration/james-database.properties
@@ -48,6 +48,6 @@ openjpa.streaming=false
 # datasource.maxTotal=8
 
 # Attachment storage
-# *WARNING*: Is not made to store large binary content
+# *WARNING*: Is not made to store large binary content (no more than 1 GB of data)
 # Optional, Allowed values are: true, false, defaults to false
 # attachmentStorage.enabled=false
