+++ server/apps/spring-app/src/main/resources/james-database.properties
@@ -46,7 +46,7 @@ openjpa.streaming=false
 # datasource.validationQuery=select 1
 
 # Attachment storage
-# *WARNING*: Is not made to store large binary content
+# *WARNING*: Is not made to store large binary content (no more than 1 GB of data)
 # Optional, Allowed values are: true, false, defaults to false
 # attachmentStorage.enabled=false
