+++ server/container/guice/jpa-common/src/test/java/org/apache/james/TestJPAConfigurationModuleWithSqlValidation.java
@@ -26,7 +26,7 @@ import java.sql.SQLException;
 
 import javax.inject.Singleton;
 
-import org.apache.james.modules.data.JPAConfiguration;
+import org.apache.james.backends.jpa.JPAConfiguration;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
